UserTagCreation/requirement.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
